system likes via js api

+++ MainApp/views.py
@@ -13,6 +13,7 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from MainApp.signals import snippet_view
 import logging
+import json
 from datetime import datetime
 
 logger = logging.getLogger(__name__)
@@ -304,25 +305,57 @@ def is_authenticated(request):
         return JsonResponse({'is_authenticated': False})
 
 
-@login_required
-def comment_like(request, id, vote):
-    comment = get_object_or_404(Comment, id=id)
+# @login_required
+# def comment_like(request, id, vote):
+#     comment = get_object_or_404(Comment, id=id)
+#
+#     # Если нет лайка/дизлайка, то мы его создаем
+#     existing_vote, created = LikeDislike.objects.get_or_create(
+#         user=request.user,
+#         content_type=ContentType.objects.get_for_model(comment),
+#         object_id=comment.id,
+#         defaults={'vote': vote}
+#     )
+#
+#     if not created:  # снимаем наш голос
+#         if existing_vote.vote == vote:
+#             # Если стоит лайк, а мы хотим убрать его, то удаляем.
+#             existing_vote.delete()
+#         else:  # меняем голос на противоположный
+#             # Если стоит лайк, а мы хотим создать дизлайк, то лайк удаляем, дизлайк создаем
+#             existing_vote.vote = vote
+#             existing_vote.save()  # -> UPDATE
+#
+#     return redirect('snippet-detail', id=comment.snippet.id)
 
-    # Если нет лайка/дизлайка, то мы его создаем
-    existing_vote, created = LikeDislike.objects.get_or_create(
-        user=request.user,
-        content_type=ContentType.objects.get_for_model(comment),
-        object_id=comment.id,
-        defaults={'vote': vote}
-    )
 
-    if not created:  # снимаем наш голос
-        if existing_vote.vote == vote:
-            # Если стоит лайк, а мы хотим убрать его, то удаляем.
-            existing_vote.delete()
-        else:  # меняем голос на противоположный
-            # Если стоит лайк, а мы хотим создать дизлайк, то лайк удаляем, дизлайк создаем
-            existing_vote.vote = vote
-            existing_vote.save()  # -> UPDATE
+def add_commen_like(request):
+    if request.method == 'POST':
+        data = json.loads(request.body)
+        comment_id = data.get("comment_id")
+        vote = data.get("vote")
 
-    return redirect('snippet-detail', id=comment.snippet.id)
+        existing_vote, created = LikeDislike.objects.get_or_create(
+            user=request.user,
+            content_type=ContentType.objects.get_for_model(Comment),
+            object_id=comment_id,
+            defaults={'vote': vote}
+        )
+
+        if not created:  # снимаем наш голос
+            if existing_vote.vote == vote:
+                # Если стоит лайк, а мы хотим убрать его, то удаляем.
+                existing_vote.delete()
+            else:  # меняем голос на противоположный
+                # Если стоит лайк, а мы хотим создать дизлайк, то лайк удаляем, дизлайк создаем
+                existing_vote.vote = vote
+                existing_vote.save()  # -> UPDATE
+
+        comment = Comment.objects.get(id=comment_id)
+        response_data = {
+            "success": True,
+            "likes_count": comment.likes_count(),
+            "dislikes_count": comment.dislikes_count(),
+        }
+
+        return JsonResponse(response_data)
